refactor(Views Mixins): create a separate package for mixins
