+++ api/views/mixins/__init__.py
@@ -1,0 +1,7 @@
+from api.views.mixins.common import ReadWriteSerializersMixin
+from api.views.mixins.group import CreateGroupMixin
+
+__all__ = [
+    "ReadWriteSerializersMixin",
+    "CreateGroupMixin",
+]
